Going back to simple logging
I tried to add too much logic in this layer of the app->server communication.

I'll go back to simple logging of every item, only avoiding duplicate id+revision

Also, I still need to publish id+revision so the server knows it arrived.  This time, it expects to see all of the revisions.